[utils] Add temporary shim for logging
Related: #5680, #7517
